+++ coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/config/LettuceConfiguration.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JavaType
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.type.TypeFactory
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import coupon_platform.infrastructure.redis.util.IssuedCouponForRedis
+import coupon_platform.infrastructure.cache.global_redis.util.IssuedCouponForRedis
 import org.springframework.boot.autoconfigure.data.redis.RedisProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
